fix gc job
gc job failed to run when it checks there is stale testbed, this is because
it always checks if the testbed is in directory `terraform.tfstate.d`, this works
when we deploy testbed and destroy testbed upon job finishes. But gc job does not
know its original directory, so we store it in /var/lib/jenkins/, the terraform
directory should be specified when calling destroy.sh.

Signed-off-by: Bin Liu <biliu@vmware.com>