Sync with cloud feature hardening - 2.
Fixes:
- Add empty string check for sg rule description field while creating sync content.
- Stale AT SG doesn't get deleted during sync as sync doesn't identify
it as an ASG associated with NSG. It is due to the usage of ASGs in "tags" field
of NSG for retrieving ASGs from cloud. User added stale ASG can't be present in tags.
Fix is to use ASGs attached to network interfaces instead of getting ASGs from "tags".
AT SGs are added as "tags" in a NSG for user visibility purpose.
- SG deletion logic uses RG name from Vnet of CloudResource. If SG belongs to
a different RG then it SG doesn't get deleted. Hence, while creating
sync content make sure that only SGs with RGs same as Vnet are included.

Signed-off-by: Archana Holla <harchana@vmware.com>