Avoid using resourceGraph query for interface data
It was been observed resource graph query doesnt give latest
data of the network interface. For example, if ASG is getting attached
on the interface and immediately in the next call, we try to update the
interface with some other ASG, that interface may or may not have the
previous ASG, leading to misleading update.

Here we use ResourceGraph query only to give interface to vnet mapping
which is used for filtering in getInterfaces call.

Signed-off-by: Rahul Jain <rahulj@vmware.com>