Add option to replay an account and selector.
Add structs to capture acccount configuration and selector
configuration.

Account config helps to know whether an account is succesfully
initialized or not, whethere credentials are valid or not. Also
provides an option to retry errors when an cloud plug-in returns
error.
Selector config keeps the latest version of selector.

This data, helps to replay all the selectors for a given account, when
account transitions for an non-initialized state to initialized
state, so that filters in cloud plug-in are re-configured with
latest version of selector.

Also move provider type to account config

Signed-off-by: Anand Kumar <kumaranand@vmware.com>