Handle ATGroup delete error and re-add.
Consider a scenario where ANP contains two rule level
AppliedToGroups. Delete one of the AppliedToGroup.
If the ATGroup could not deleted on the cloud, then the
internal state of network policy controller still contains
the rules, while the rules are deleted on the cloud.

This PR tries addresses the issue, by clearing the
internal state when an ATGroup is deleted, so that
if the ATGroup is re-added, then NP controller will be
able to recover and realize the rules on the cloud.

Signed-off-by: Anand Kumar <kumaranand@vmware.com>