Track appliedTo group to network policies
Tracker is used to update NP Status on Virtual Machine Policy
Object. In an event ANP is deleted and cloud resources cannot be
cleaned up, the ANP information is lost as its deleted from indexer
leading to misleading info in vmp cache.

With this PR, we store the mapping of appliedTo to Network Policy
and use it for cases when ANP is not available in the cache.

Signed-off-by: Rahul Jain <rahulj@vmware.com>