Validate interface to object conversion
When the ByIndex function of cache indexer is used, it returns a list of
interfaces. Validate if the interface can be converted to an object, before
accessing the actual object.

When traversing through a list of interfaces, It is observed that the
interface can be nil.
If not handled, then it will result in a crash with error saying
"panic: interface conversion: interface {} is nil, not *XXXX"

Signed-off-by: Anand Kumar <kumaranand@vmware.com>