Update retry queue processing
Retry queue was processed every second and the retry count was 15.
So, for a failed cloud operation, it was retrying very fast and too many
number of times.

With this PR, we are reducing the number of retry to 3 and increased the
retry interval to 20 sec. So essentially a failed cloud response will be
retried for 1 minute with 20 sec gap.

Signed-off-by: Rahul Jain <rahulj@vmware.com>