Avoid retrying while syncing security groups
Nephe runs periodic sync with cloud routine to sync cloud entity
with internal cache. If any error occurs while syncing cloud entities,
we want to avoid adding the entity in retry queue. This is to avoid
unneccessary retries if sync with cloud fails, as it itself is another
retry mechanism.

With this PR, we mantain a new state on cloud entity to indicate if
that entity is eligibile for retry if any error occurs.

Signed-off-by: Rahul Jain <rahulj@vmware.com>