Support deny action for Azure cloud (#312)
With this PR, Nephe will support Deny action for Azure cloud. If an
ANP is configured with Drop action, it will be translated to Deny
action for Azure.
Nephe will also support ANP tier and priority and re-arrange the rules
based on priority. Lower the priority, higher the precedence.

Changes
- Introduce priroity and action in CloudRule
- Translate ANP tier, priority, and rule priority to CloudRule priority
- Cloud Rule will store rule priority in the rule description field which
  will be used to insert new rule at its correct location/priority
- Update sync logic to ignore action/priority for aws cloud
- Update logic to detect ANP modify, in case of tier or priority changes
- Update unit tests
- Add integration test for validating priority re-ordering

Signed-off-by: Rahul Jain <rahulj@vmware.com>